Feature/finished/ikm kec kl2 (#196)
* Update version to start feature: IKM-KEC-KL2

* Clarify intent of factory field interface and implementation.

The FieldInterface provides the interface contract, and the FieldImplementation provides the concrete class that the factory will construct in case that is needed for any reason. Possibly the concrete class or the factory class may be stored in preferences to reconstruct a particular layout.

* Added additional factory interface contracts and simplified naming.

Moving beyond fields to Factories for Tinkar components (Patteren, Concept, Semantic, Stamp, Version). Shortened naming scheme, moved muti-component/version layouts to their own package. Deprecated classes for removal soonish.

* Update version to finish feature: IKM-KEC-KL2