JPRO-169-171 Allow statechanges of Property-{Concept/Pattern/Semantic}-Windows to persiste (#756)
Since this commit touches the preference backing system, it will only work on journalwindows that once got its panes recreated.

- fix bug that did not allow custom prop's to be persistate if there type was none boolean

- fix small fxml bug in Pattern/PatternSematics FXML around CommentsButton

- property patterns are currently saved with custom STRINGS representing there state
In a later PR this can be more generaly hocked into ModelView state of Windows.

Notice that this commit allows for statechanges to persist through exeting and reopening the respective Journal.
On the Pattern window for example closing and opening the property tab will **not** remeber that state as this is
handeld internally in teh window itself. Once the situation is rectified there the solution can directly be hoocked up
into this solution.

Co-authored-by: Carl Dea <carl.dea@gmail.com>