JPRO-200 allow for state changes in FilterMenuPopup to persist
Implementing a ViewModel to detach some of the pathMap state into the ViewModel

This can further be utilized to also detach the TimelineController from more state, like its done in the new FilterMenuController.

This PR also added the bonus feature that any changes in the FilterMenu are now instantly updated in the Timeline itself.