JPRO-200 allow for state changes in FilterMenuPopup to persist (#768)
* JPRO-200 allow for state changes in FilterMenuPopup to persist

Implementing a ViewModel to detach some of the pathMap state into the ViewModel

This can further be utilized to also detach the TimelineController from more state, like its done in the new FilterMenuController.

This PR also added the bonus feature that any changes in the FilterMenu are now instantly updated in the Timeline itself.

* fix race where tracked moduleNIDS were not in line with path updates

* add missing licence

---------

Co-authored-by: Carl Dea <carl.dea@gmail.com>