JPRO-210 correctly save changed EntityFacade to preference system. (#808)
This does not only fix it for the Pattern window, but also for Concept and PatternSemantics/GenEditing in a general way.

The presented way of saving preferences on change in the inner ViewModel should be seen as the prefered way of updating state persistenly.
In future work this code could be expaned to also handle saving preferences in a central clean way on ViewModel changes such as propertPanel state.

Co-authored-by: Carl Dea <carl.dea@gmail.com>