JPRO-206 fix description cutout on concept window rerender
This bug service through a race between minHeight and prefHeight calculations, where under rerender conditions a scenario is arived where the minHeight is bigger then the prefHeight when the prefHeight
Skin ovverride is executed, confusing the render enginge.

Since in layouting this particular Node we only care about prefHeight/Width we shall not recalulating the minHeight. This seams to not have any negative visual impacts uppon testing.